Added javascript to confirm 'Delete' post.

+++ admin/includes/view_all_posts.php
@@ -112,7 +112,7 @@ if (!$select_all_posts) {
     echo "<td><a href='posts.php?publish={$post_id}'>Publish</a></td>";
     echo "<td><a href='posts.php?draft={$post_id}'>Draft</a></td>";
     echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-    echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+    echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?');\" href='posts.php?delete={$post_id}'>Delete</a></td>";
     echo "</tr>";
   };
 }
